Added mipmaps, fixed rotation bug, fixed drawing of stars twice
A gluBuild2DMipmaps is now used to build texture with mipmaps.

A rotation bug was a situation when cursor jumped up/down each time you
wanted to rotate above the upper limit -- it was due to the value of the old
rotation was validated rather then value that was supposed to be assigned to
it (new value).

And as of starse, one of the parameters should be in a domain [0, 180]
whereas in the old code both were in [0, 360].