Makefile changed to generate sgi textures automatically from jpg files.
So far, the sgi images used by the programs as textures were kept with the
rest of the source files and data in dist folder was a symbolic link to the
data directory in root source directory.  After this change SGI images are
removed from source distribution and automatically generated from JPG files.