gl::AbstractObjects -> gl::AbstractScene
This commit replaces the usage of term "objects" in favour of
"scene" so we have now gl::AbstractScene and physics::Scene rather
than gl::AbstractObjects and physics::Objects.  This should be
less confusing.

Also, gl::AbstractScene::ptr has been added which is just a
typefed for std::auto_ptr<>.  This should coding easier.