Major refactorisation of files and introducing of load/save methods
This commit moves around all the files dividing them in more
logical directories.  There is a lib directory for library
routines (sintable, constants, lexer), an ui directory for
user interface elements (widgets, windows; with notable
exception of gl::Widget) and gl directory for OpenGL helper
classes.

This commit also introduces gl::AbstractScene::load() and
gl::AbstractScene::save() methods (the former being virtual,
the latter being static) which will be used to load and save
scenes from files.

The initialize() function has been renamed to
createMainWindow() and it's role now is to merely create
a main application window.

The scene is not loaded when program starts now and instead
a "Load" menu field will be introduced for that purpose.