graph::Graph implemented
The graph::Graph class represents an undirected graph with no
(v, v) edges.  Edges are stored in an "optimised" matrix meaning
that there is only one entry for (u, v) and (v, u) edges (since they
are the same) plus there are no entries for (v, v) edge.

Example code of how to traverse edges is provided as a comment.