Fix segfault on run and change ownership rules a little
Previously, the application segfaulted on run.  This was most
likely caused by the pane not being initialised in MainWindow
constructor.  This commit changes fixes that and changes ownership
rules a bit.  GLPane is now constructed by MainWindow() so
there are no problems with pane being NULL.

This commit also moves file dialog handling to a single, static
function instead of having similar code in both load() and save()
functions.

Finally, this commit removes init.cpp file incorporating it with
the mainwindow.cpp.