Created abstract configuration class
Created a set of class for defining various types of configuration.
The classes have a facility of run-time values discovery so that
code that handles configurations may be written in generic manner.

The ui::cfg::Data class is a base class for classes with
configuration values.  This class stores a list of pointers to
Value objects provided by a class that inherit from it.  This
makes it possible to detect at run-time what values the
configuration object contains.

The ui::Config<Data> template lets one safely pass the configuration
data around as it implements a shared pointer mechanism.