Fixed problem with initializeGL() never being called
The gl::AbstractScene::initializeGL() was called in
gl::Widget::initializeGL() but the latter is called only
once when widget is created.  At this point, no scene is
attached to the widget.  This means that scene's initializeGL()
is never called.

This commit moves call to initializeGL() to paintGL() so that
gl::AbstractScene::initializeGL() is called on each frame.  This
is not a preferred solution but works for now.