Added first version of evolutionary solver.
This isn't really an evolutionary solver and is subject to further
changes. This version allows for only one iteration in fixed conditions.
Furthermore, after this one iteration all pre-existing population is
forgotten (well, to be honest, it's worse -- it's ignored). The best
solution is inserted into scene, and viewport is redrawn.

Additionally, I had to add gsl-related include to rand.hpp.