evo: various tweaks
The Score class has been removed and instead of "valid" flag
a quiet NaN is used to distinguish invalid scores.  Also, instead
of storing graph and it's score in a std::pair<> an Individual
class has been created with definition in .cpp file.

Definition of the Compare class has also been moved to .cpp file.

The evaluate() method has been moved outside from the Compare
class and defined as global function in the cpp file.